Refactored associated GlucoseMeasurement types into their own files
This is, because we might need to add more types. Well not might, we will.
